backend/requirements.txt
pandas
transformers
torch>=2.4.0
google-genai
python-dotenv